requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pyyaml==6.0
requests==2.31.0
sqlalchemy==2.0.23
python-multipart==0.0.6
jinja2==3.1.2
python-dateutil==2.8.2
dnspython==2.4.2
aiohttp==3.8.5
pydantic==1.10.12
aiodns==3.1.1
pycares==4.4.0
cloudpickle==3.0.0
